Bugfix: Crouch-spamming midair jump
BUG: Crouch spamming in air can cause you to raise up in the air.
Would've been cool as a feature, except if it wasn't letting you jump, somehow, possibly due to async of the boxcast and the hitbox.
(HARD) FIX: Jesus Himself now forbids any uncleansed soul to jump while holsing the S button, A.K.A and FYI - crouching.
But seriously... that's it. No more jumping while crouching.